requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
boto3==1.14.49
    # via -r requirements.in
botocore==1.17.63
    # via
    #   boto3
    #   s3transfer
build==0.8.0
    # via pip-tools
click==8.1.3
    # via pip-tools
cssselect2==0.7.0
    # via svglib
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-js-asset
    #   django-polymorphic
    #   django-sekizai
    #   django-select2
    #   django-storages
    #   django-treebeard
    #   djangocms-frontend
    #   easy-thumbnails
django-admin-sortable2==1.0.4
django-appconf==1.0.5
    # via django-select2
django-classy-tags==3.0.1
    # via
    #   -r requirements.in
    #   django-cms
    #   django-sekizai
django-cms==3.11.0
    # via
    #   -r requirements.in
    #   djangocms-attributes-field
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-googlemap
    #   djangocms-icon
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-text-ckeditor
    #   djangocms-video
django-entangled==0.4
    # via djangocms-frontend
django-filer==2.2.3
    # via
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-googlemap
    #   djangocms-picture
    #   djangocms-video
django-formtools==2.3
    # via django-cms
django-js-asset==2.0.0
    # via django-mptt
django-mptt==0.13.4
    # via django-filer
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.0.0
    # via
    #   -r requirements.in
    #   django-cms
django-select2==7.10.1
    # via djangocms-frontend
django-storage-url==0.5.0
    # via -r requirements.in
django-storages==1.13.1
    # via django-storage-url
django-treebeard==4.4
    # via
    #   django-cms
    #   djangocms-attributes-field
djangocms-admin-style==3.2.0
    # via
    #   -r requirements.in
    #   django-cms
djangocms-attributes-field==2.1.0
    # via
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-icon
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-video
djangocms-file==3.0.0
    # via -r requirements.in
djangocms-frontend==1.0.0
    # via -r requirements.in
django_extensions==3.2.1
djangocms-googlemap==2.0.0
    # via -r requirements.in
djangocms-icon==2.0.0
    # via -r requirements.in
djangocms-picture==3.0.0
    # via -r requirements.in
djangocms-style==3.0.0
    # via -r requirements.in
djangocms-text-ckeditor==5.1.1
    # via
    #   -r requirements.in
    #   djangocms-frontend
djangocms-video==3.0.0
    # via -r requirements.in
docutils==0.15.2
    # via botocore
easy-thumbnails[svg]==2.8.3
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-picture
environs==9.5.0
furl==2.1.3
    # via django-storage-url
html5lib==1.1
    # via djangocms-text-ckeditor
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
lxml==4.9.1
    # via svglib
orderedmultidict==1.0.1
    # via furl
packaging==21.3
    # via build
pep517==0.13.0
    # via build
pillow==9.2.0
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
parler==1.0.1
pip-tools==6.8.0
    # via -r requirements.in
psycopg2==2.8.5
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via botocore
pytz==2022.2.1
    # via
    #   -r requirements.in
    #   django
reportlab==3.6.11
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.3.7
    # via boto3
six==1.16.0
    # via
    #   furl
    #   html5lib
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.4.2
    # via django
svglib==1.4.1
    # via easy-thumbnails
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via
    #   build
    #   pep517
unidecode==1.1.2
    # via django-filer
urllib3==1.25.11
    # via botocore
uwsgi==2.0.19.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.37.1
    # via pip-tools
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools